refactor(dsp): remove remaining module-level DSP implementations, delegate to @wake-studio/dsp
- afe-aec/afe-bss/rnnoise AFE-stage: replace duplicated local levelDb with
  @wake-studio/dsp import (3 copies deleted).
- rnnoise core/dsp: frameRms/applyGain now re-export from dsp package.
- few-shot core/dsp: peakDbfs/rmsDbfs/isClipped/estimateSnrDb moved to
  @wake-studio/dsp level.ts; few-shot keeps only cosine similarity + quality
  assembly.
- dsp level.ts gains peakDbfs/rmsDbfs/isClipped/frameRms/applyGain/
  estimateSnrDb.
- Remaining in-module files named dsp.ts contain domain logic (KWS
  smoothing/trigger/VAD gate), not numeric DSP - left in place.
- Full-repo typecheck/lint/unit green; web build bundles fft.js in worklet.
