+++ packages/modules/few-shot/package.json
@@ -19,6 +19,7 @@
   },
   "dependencies": {
     "@wake-studio/contracts": "workspace:*",
+    "@wake-studio/dsp": "workspace:*",
     "@wake-studio/module-kws-engine": "workspace:^",
     "@wake-studio/module-kws-plix": "workspace:^",
     "@wake-studio/platform": "workspace:^"
